PYTHON_RECOMMENDATION_ENGINE/recommedation_web_crawler/news_crawler.py
from GoogleNews import GoogleNews
news = GoogleNews(period='1d')
news.search("India")
result = news.result()
import pandas as pd
data = pd.DataFrame.from_dict(result)
data = data.drop(columns=["img"])
print(data.head)