PYTHON_RECOMMENDATION_ENGINE/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
APScheduler==3.10.4
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lxml==5.2.1
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.1
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
threadpoolctl==3.4.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
